// src/app/domain/modules/cartera/conciliacion/conciliacion.interfaces.ts
export interface ConciliacionI {
  Anio: string;
  Dias: number;
  Estado: string;
  FechaConciliacion: string;
  IdConciliacion: number;
  IdGestion: number;
  Identificacion: string;
  Mes: number;
  NActaConciliacion: string;
  Nit: string;
  RutaArchivo: string;
  Tercero: string;
  Trimestre: number;
  UsuarioEncargado: string;
  ValorCancelado: number;
  ValorConciliado: number;
  ValorCuotaModeradora: number;
  ValorDeGlosasAceptadaPorIPS: number;
  ValorDevuelto: number;
  ValorDiferencia: number;
  ValorEnAuditoria: number;
  ValorEnRetencion: number;
  ValorGlosado: number;
  ValorNoDescontadoPorEps: number;
  ValorNoRadicado: number;
  ValorPagoNoAplicado: number;
  ValorReconocidoParaPago: number;
  ValorTotal: number;
}